PRでご指摘いただいた部分の修正
