+++ src/core/domain/tagdm/tag_repository.go
@@ -4,5 +4,5 @@ import "context"
 
 type TagRepository interface {
 	FindById(ctx context.Context, id TagId) (*Tag, error)
-	FindByTagName(ctx context.Context, tagName TagName) (*TagId, error)
+	FindIdByTagName(ctx context.Context, tagName TagName) (*TagId, error)
 }
