+++ src/core/domain/userdm/self_introduction.go
@@ -1,6 +1,10 @@
 package userdm
 
-import "github.com/cockroachdb/errors"
+import (
+	"unicode/utf8"
+
+	"github.com/cockroachdb/errors"
+)
 
 type SelfIntroduction string
 
@@ -9,7 +13,7 @@ func NewSelfIntroduction(value string) (*SelfIntroduction, error) {
 		return nil, errors.New("SelfIntroduction is empty")
 	}
 
-	if len(value) > 2000 {
+	if utf8.RuneCountInString(value) > 2000 {
 		return nil, errors.New("SelfIntroduction is too long")
 	}
