+++ src/core/infra/controllers/create_user_controller.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/TakayukiHirano117/architecture-study/src/core/app/userapp"
-	"github.com/TakayukiHirano117/architecture-study/src/core/domain/domain_service"
+	"github.com/TakayukiHirano117/architecture-study/src/core/domain/tagdm"
 	"github.com/TakayukiHirano117/architecture-study/src/core/domain/userdm"
 	"github.com/TakayukiHirano117/architecture-study/src/core/infra/rdbimpl"
 	"github.com/gin-gonic/gin"
@@ -12,15 +12,15 @@ import (
 
 type CreateUserController struct {
 	userRepo          userdm.UserRepository
-	userDomainService domain_service.UserDomainService
-	tagDomainService  domain_service.TagDomainService
+	userDomainService userdm.UserDomainService
+	tagDomainService  tagdm.TagDomainService
 }
 
 func NewCreateUserController() *CreateUserController {
 	return &CreateUserController{
 		userRepo:          rdbimpl.NewUserRepositoryImpl(),
-		userDomainService: domain_service.NewUserDomainService(rdbimpl.NewUserRepositoryImpl()),
-		tagDomainService:  domain_service.NewTagDomainService(rdbimpl.NewTagRepositoryImpl()),
+		userDomainService: userdm.NewUserDomainService(rdbimpl.NewUserRepositoryImpl()),
+		tagDomainService:  tagdm.NewTagDomainService(rdbimpl.NewTagRepositoryImpl()),
 	}
 }
